implement tl status tracker in traffic_light_compliance_checker library
- add new class TrafficLightStatusTracker to compliance checker library
- move signal history logic from traffic_light_filter to TrafficLightStatusTracker
